requirements.txt
pywin32
docker-py